update usages of MethodTween and its subclasses to reflect change from float to double

+++ src/core/Music/Playlist.cs
@@ -306,7 +306,7 @@ namespace UTheCat.Jumpvalley.Core.Music
                         FinalValue = 1
                     };
 
-                    currentTween.OnStep += (object o, float _frac) =>
+                    currentTween.OnStep += (object o, double _frac) =>
                     {
                         SceneTreeTween t = (SceneTreeTween)o;
                         SetLinearVolumeViaTween(t.GetCurrentValue());
